Load viz.js from URL given in script tag instead of hard-coded
The type="javascript/worker" attribute in the script tag doesn't
really mean anything in itself, but since it's not "text/javascript"
or "application/javascript" it causes the browser to just load the
script, but not parse it. That's what we want, since the script will
be parsed anyway when the web worker imports it.

The purpose is twofold:

1. Pre-fetch: Start loading it directly when the page loads in order
   to have (at least most of) it in memory when the worker imports it

2. Transparency: A clear way to specify from which URL to load viz.js