Removed unused stuff
This particular stuff is unused since shape tweening by converting
some shapes to paths is only supported for polygons and ellipses,
which are the only shapes that graphviz is known to generate. This
means that the convertToPathData function is called only for polygons
and ellipses and it is enough to check if it is a polygon and assume
it's an ellipse otherwise.