Support getting viz URL from normal script tag
Avoids the need to have multiple script tags when using viz.js both in
worker and directly in drawNode().

If DrawNode() is not used, the startup time can be made shorter if a
script tag of type "javascript/worker" is used, since JIT compilation
does not need to be done twice.