Create a new document before each sub-test in transition test
The test is not really built for anything else. Although it works as
it is now, it will not work when an existing graphviz renderer is
reused. One reason is that the test terminates the worker after each
sub-test and the graphviz renderer will not create a new one when it
is reused. Another reason is that each sub-test assumes that no
previous graph exists the first time a graph is rendered, which is not
the case when the graphviz renderer is reused.