Fix undefined bug: Only set datum.offset if prevStartShape && startShape are defined
If the shapes tags doesn't match any of the values "polygon", "ellipse" or "path",
which can happen if you have a node with e.g. shape=none, then prevStartShape and/or startShape
will be undefined and an error will be thrown due to trying to get .center.

Fix by checking if both are defined, and then setting datum.offset.