Fix error when growing edge from shape none or plain
These shapes have no svg shape so we use the label position of the
start shape and previous start shape to calculate the start offset for
the growing edge.

If the node has no label we use a zero offset. This doesn't make the
edge grow from the correct place, but since it's invisible, it may be
acceptable anyway.

Fixes https://github.com/magjac/d3-graphviz/issues/103.