Fix styles removed after transition
The opacity style property is added and used to fade in or out during
transitions. After the transition it is should be removed. Instead of
doing that by removing the style attribute, restore it to the
original.

Fixes https://github.com/magjac/d3-graphviz/issues/159.

squash! Fix styles removed after transition