remove now useless deletion of hpccWasm.graphviz in tests
Before moving to ESM modules this was used to test the code in dot.js
which avoids JIT compilation of the @hpcc-js/wasm when a web worker is
used. This hack no longer works and instead the code in dot.js that
cannot be executed in test has been excluded form coverage.