remove useless deletion of hpccWasm.graphviz in dot-shared-worker-test.js
This was used to test the code in dot.js which avoids JIT compilation
of the @hpcc-js/wasm when a web worker is used. This hack will no
longer works when moving to ES modules and instead the code in dot.js
that cannot be executed in test has been excluded from coverage.