temporarily disable tests that will fail with Graphviz upgrade
These tests would otherwise fail when Graphviz is upgraded to version
8.1.0 in the next commit in this series. The reason is that the
Graphviz issue "Failure of arrowhead and arrowtail to respect
penwidth" (https://gitlab.com/graphviz/graphviz/-/issues/372) was
fixed for the use cases in these tests in version 8.1.0 so the data
has slightly changed. The tests will be updated and re-enabled in
separate upcoming commits in this series.