Port codecov to use Codecov's Github Action dedicated wrapper
The immediate reason for doing this is that running codecov failed
with:

```
==> Uploading reports
/home/magjac/d3-graphviz/node_modules/codecov/lib/codecov.js:217
      } else if (result.split('\n').length !== 2) {
                        ^

TypeError: result.split is not a function
    at /home/magjac/d3-graphviz/node_modules/codecov/lib/codecov.js:217:25
    at /home/magjac/d3-graphviz/node_modules/teeny-request/build/src/index.js:217:17
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)

Node.js v18.19.1
```

The Codecov NodeJS Uploader¹ has been deprecated and seems to no
longer work. Instead, uploading reports to Codecov using the Codecov
CLI² is now recommeded. This change does this using the Codecov GitHub
Action³.

¹ https://github.com/codecov/codecov-node
² https://docs.codecov.com/docs/codecov-uploader
³ https://github.com/marketplace/actions/codecov